test(fix): refactor to fix tests
